pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
]
